django-app/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pytz==2024.1
setuptools==69.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
wheel==0.42.0
